+++ packages/config/src/Config.ts
@@ -1,0 +1,7 @@
+import type { Type } from ".";
+
+export class Config {
+  public static make(): Type. {}
+
+
+}
