+++ packages/config/src/types/index.ts
@@ -1,2 +1,2 @@
-export * from './global';
+export * from './window';
 export * from './types';
