+++ packages/config/src/types/types.ts
@@ -34,55 +34,24 @@ export module Type {
 		name: string;
 		offset: number;
 	}
-}
 
-export type Api = {
-	graphqlEndpoint: string;
-	restApiBaseUrl: string;
-	restApiCollectionEndpoints: Record<string, string>;
-	restApiNonce: string;
-	restApiPrimaryKeys: Record<string, string | [string]>;
-	restApiRouteUrl: string;
-};
+	export interface GeneralSettings {
+		dateFormat: string;
+		timeFormat: string;
+		timezone: string;
+	}
 
-export type Config = {
-	brandName: string;
-	currency: Type.Currency;
-	currentUser: User; // LATER: consolidate data types
-	dateTimeFormats: Type.DateTimeFormats;
-	generalSettings: GeneralSettings;
-	locale: Type.Locale;
-	nonce: string;
-	sitePermissions?: string[]; // LATER: consolidate data types
-	siteUrl: Type.SiteUrls;
-	timezone: Type.Timezone;
-	wp_debug: boolean;
-};
-
-export interface GeneralSettings {
-	dateFormat: string;
-	timeFormat: string;
-	timezone: string;
-}
-
-export type CoreDomainDomData = {
-	assetNamespace: string;
-	brandName: string;
-	coreVersion: string;
-	distributionAssetsPath: string;
-	distributionAssetsUrl: string;
-	pluginPath: string;
-	pluginUrl: string;
-};
-
-export interface LocaleDomData {
-	user: string;
-	siteTimezone: LocaleTimezoneDomData;
-	site: string;
-}
-
-export interface LocaleTimezoneDomData {
-	city: string;
-	name: string;
-	offset: number;
+	export interface Config {
+		brandName: string;
+		currency: Currency;
+		currentUser: User; // LATER: consolidate data types
+		dateTimeFormats: DateTimeFormats;
+		generalSettings: GeneralSettings;
+		locale: Locale;
+		nonce: string;
+		sitePermissions?: string[]; // LATER: consolidate data types
+		siteUrl: SiteUrls;
+		timezone: Timezone;
+		wp_debug: boolean;
+	}
 }
